api-service/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
drf-spectacular
dj-database-url
django-filter
gunicorn
# ML components from previous setup
prophet
pandas
numpy
reportlab
